move diagnostics to a separate module
