+++ pkg/diagnostics/server.go
@@ -1,4 +1,4 @@
-package server
+package diagnostics
 
 import (
 	"context"
