+++ pkg/xds/server/components.go
@@ -32,21 +32,7 @@ var (
 	xdsServerLog = core.Log.WithName("xds-server")
 )
 
-func SetupDiagnosticsServer(rt core_runtime.Runtime) error {
-	return rt.Add(
-		// diagnostics server
-		&diagnosticsServer{
-			port:    rt.Config().XdsServer.DiagnosticsPort,
-			metrics: rt.Metrics(),
-		},
-	)
-}
-
 func SetupServer(rt core_runtime.Runtime) error {
-	err := SetupDiagnosticsServer(rt)
-	if err != nil {
-		return err
-	}
 	reconciler := DefaultReconciler(rt)
 
 	metadataTracker := NewDataplaneMetadataTracker()
